Initial conversion to markdown format.
I've included a python script in tools/html-to-md.py
that I used to do this conversion, in case anyone
wants to reconstruct it. It does a rudimentary job,
slightly tailored to the structure of the existing
DWARF HTML pages.

After the conversion, I did some manual tweaks, and
rearranged a few pages.